fix: 상대경로 수정
