+++ ai-web-back/api/tasks.py
@@ -1,15 +1,10 @@
 from myapp.celery import app
-import time
 from .models import Review
 
-
-import os, sys
-sys.path.append(os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__)))))
-##from ai_web_pytorch.review_model import ReviewService
-from ai_web_pytorch import model
+from .ai_web_pytorch.review_model import ReviewService
 
 
-##model = ReviewService()
+model = ReviewService(model_path='./api/ai_web_pytorch/model.pth')
 
 @app.task()
 def predict(string:str, review_id:int) -> None:
