--- ai_web_pytorch/__init__.py
@@ -1,3 +1,0 @@
-from .review_model import ReviewService
-
-model = ReviewService()
